Move OudsGridKeyToken file in tokens directory
